internal - move JiT helpers only needed by SYCL to ceed/jit-tools-deprecated.h; this change should have no public facing effects
