+++ backends/sycl-shared/ceed-sycl-shared-basis.sycl.cpp
@@ -7,6 +7,7 @@
 
 #include <ceed/backend.h>
 #include <ceed/ceed.h>
+#include <ceed/jit-tools-deprecated.h>
 #include <ceed/jit-tools.h>
 
 #include <map>
